Apply our new best practices
