+++ .githooks/pre-commit/test
@@ -5,7 +5,6 @@
 #
 
 PATCH_FILE="working-tree.patch"
-NPM_BIN="./node_modules/.bin"
 
 function cleanup {
     exit_code=$?
@@ -24,5 +23,5 @@ git checkout -- .
 
 git_cached_files=$(git diff --cached --name-only --diff-filter=ACMR | xargs echo)
 if [ "$git_cached_files" ]; then
-    npm run validate || exit 1
+    npm run validate --silent || exit 1
 fi
