+++ configs/development/hosts.json
@@ -1,0 +1,3 @@
+{
+    "static": "/"
+}
