--- configs/project.js
@@ -1,13 +1,0 @@
-module.exports = {
-    i18n: {
-        langs: ['en', 'ru']
-    },
-
-    supportedBrowsers: [
-        'IE >= 9',
-        'Safari >= 5',
-        'Chrome >= 33',
-        'Opera >= 12.16',
-        'Firefox >= 28'
-    ]
-};
