--- configs/testing/env.js
@@ -1,7 +1,0 @@
-module.exports = {
-    socket: '/tmp/yandex-maps-ui-{{ProjectName}}.socket',
-    cluster: {
-        workers: 2,
-        timeout: 60
-    }
-};
