+++ configs/testing/hosts.json
@@ -1,0 +1,3 @@
+{
+    "static": "//yastatic.net/{{ProjectName}}/{{DEBIAN_VERSION}}/"
+}
