+++ package.json
@@ -38,10 +38,10 @@
     "ym": "^0.0.8"
   },
   "scripts": {
-    "start": "DEBUG={{ProjectName}}:* PORT=8080 supervisor -w server,configs -- server/boot.js",
-    "validate": "npm run lint && npm test",
+    "start": "./start-development-server.sh",
+    "validate": "npm run lint --silent && npm test --silent",
     "lint": "jshint . && jscs .",
-    "test": "npm run test-client && npm run test-server",
+    "test": "npm run test-client --silent && npm run test-server --silent",
     "test-client": "enb make tests/client -n && mocha-phantomjs tests/client/index.html",
     "test-server": "mocha -R spec --recursive tests/server",
     "build": "enb make",
