+++ server/boot.js
@@ -1,15 +1,11 @@
-var fs = require('fs');
 var path = require('path');
 var recluster = require('recluster');
-
 var debug = require('debug')('{{ProjectName}}:boot');
-var env = require('../configs/current/env');
 
-try {
-    fs.unlinkSync(env.socket);
-} catch (error) {}
-
-var cluster = recluster(path.join(__dirname, 'app.js'), env.cluster);
+var cluster = recluster(path.join(__dirname, 'app.js'), {
+    workers: process.env.CLUSTER_WORKERS || 2,
+    timeout:  process.env.CLUSTER_TIMEOUT || 60
+});
 cluster.run();
 
 /**
